requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Browser Automation
playwright==1.40.0

# HTTP Requests
httpx==0.25.2
requests==2.31.0

# Data Processing
pandas==2.1.3
numpy==1.26.2

# PDF Processing
pypdf2==3.0.1
pdfplumber==0.10.3

# Image Processing
Pillow==10.1.0

# Visualization
matplotlib==3.8.2
plotly==5.18.0

# HTML Parsing
beautifulsoup4==4.12.2
lxml==4.9.3

# LLM Integration (Anthropic Claude)
anthropic>=0.18.0

# Utilities
python-dotenv==1.0.0